+++ NEWS.md
@@ -1,6 +1,11 @@
 
 # callr development version
 
+* callr can now pass the environment of the function to the subprocess,
+  optionally. This makes it easier to call an internal function of a
+  package in a subprocess. See the `package` argument of `r()`, `r_bg()`,
+  `r_session$run()`, etc. (#147).
+
 # callr 3.4.4
 
 * An `r_session` now exits if the load hook errors. This generates an error